NAS-129775: Put function outside service and into utils
